refactor: Product 도메인 RepositoryImpl 통합
