+++ api/src/main/java/com/albert/commerce/product/command/application/dto/ProductsResponse.java
@@ -1,4 +1,4 @@
-package com.albert.commerce.product.command.application;
+package com.albert.commerce.product.command.application.dto;
 
 import com.albert.commerce.common.units.BusinessLinks;
 import lombok.AccessLevel;
